rename 'tests/vectordb' to 'tests/vector_db'
