+++ tests/vector_db/test_chroma_db.py
