+++ tests/vector_db/test_pinecone.py
